feat: Introduce ensure-cmd.sh
To simplify the use of tmplr when building other projects, one can copy
the ensure-cmd.sh into the project and call it before building it. If
tmplr is available and the version matches, then the preinstalled
version is used. However, if tmplr is not installed, then it can be
downloaded and compiled automatically.

Signed-off-by: Diogo Behrens <db7@sdf.org>